requirements.txt
pydantic
pyyaml
mcp
anthropic